requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudscraper==1.2.52
cryptography==3.3.1
decorator==4.4.2
idna==2.10
pixiv-api==0.3.6
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==13.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
yapf==0.30.0
